data-pipeline/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.16.0
python-dotenv>=1.0.0
pymongo>=4.6.0
psycopg2-binary>=2.9.9
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0
jupyter>=1.0.0
pytest>=7.4.0
black>=23.7.0
flake8>=7.0.0
isort>=5.12.0
mypy>=1.5.0 